refactor: extract structured data util
