+++ src/components/StructuredSEO.tsx
@@ -2,42 +2,7 @@ import React from 'react'
 import { Helmet } from 'react-helmet'
 import { useTranslation } from 'react-i18next'
 import { projects } from '../data/projects'
-
-const SITE_URL = 'https://karimhammouche.com'
-
-export const getStructuredData = (lang: 'fr' | 'en') => {
-  const websiteData = {
-    '@context': 'https://schema.org',
-    '@type': 'WebSite',
-    url: SITE_URL,
-    name: 'Karim Hammouche | Portfolio',
-    inLanguage: lang,
-  }
-
-  const personData = {
-    '@context': 'https://schema.org',
-    '@type': 'Person',
-    name: 'Karim Hammouche',
-    url: SITE_URL,
-  }
-
-  const projectData = projects.map((p) => {
-    const description = Array.isArray(p.description[lang])
-      ? p.description[lang].join(' ')
-      : p.description[lang]
-
-    return {
-      '@context': 'https://schema.org',
-      '@type': 'CreativeWork',
-      name: p.title[lang],
-      description,
-      image: p.image,
-      ...(p.url ? { url: p.url } : {}),
-    }
-  })
-
-  return { websiteData, personData, projectData }
-}
+import { getStructuredData } from '../utils/structuredData'
 
 const StructuredSEO: React.FC = () => {
   const { i18n } = useTranslation()
